Added basic calculator

+++ calcu.c
@@ -6,16 +6,16 @@
 */
 int main()
 {
-	int num1;
-	int num2;
+	double num1;
+	double num2;
 
 	printf("Enter your first number: ");
-	scanf("%d", &num1);
+	scanf("%lf", &num1);
 	
 	printf("Enter your second number: ");
-	scanf("%d", &num2);
+	scanf("%lf", &num2);
 
-	printf(" Total: %d", num1 + num2);
+	printf(" Total: %lf", num1 + num2);
 
 	return 0;
 }
